feat: removed beehiiv integration
