+++ .env.local.example
@@ -13,10 +13,6 @@ NETLIFY_API_TOKEN=netlify_api_token_here
 # Exchange Rate API (exchangerate.host) - Server-side only
 EXCHANGE_RATE_API_KEY=your_access_key_here
 
-# Beehiiv Newsletter Configuration - Server-side only
-BEEHIIV_API_KEY=your_beehiiv_api_key_here
-BEEHIIV_PUBLICATION_ID=your_publication_id_here
-
 # Resend Newsletter Configuration
 RESEND_API_KEY=re_xxxxxxxxxxxxx
 RESEND_FROM_EMAIL=hello@mail.ragtechdev.com
